pyproject.toml
[project]
name = "ai-news-direct"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crawl4ai>=0.6.3",
    "python-dateutil>=2.9.0.post0",
]
